% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analizar_glmm.R
\name{analizar_glmm}
\alias{analizar_glmm}
\title{Ajustar y analizar un Modelo Lineal Mixto Generalizado (GLMM)}
\usage{
analizar_glmm(
  datos,
  formula_fijos,
  aleatorios,
  tipo = c("conteos", "presencia_ausencia", "poisson", "binomial", "negativa_binomial",
    "binomial_negativa"),
  diagnosticos = TRUE
)
}
\arguments{
\item{datos}{Un \code{data.frame} que contiene las variables del modelo.}

\item{formula_fijos}{Una formula de R o una cadena de caracteres para la parte de efectos fijos (ej. \code{y ~ x1}).}

\item{aleatorios}{Una cadena de caracteres que define la estructura de efectos aleatorios (ej. \code{"(1 | bloque)"}).}

\item{tipo}{Tipo de respuesta. Puede ser \code{"conteos"},
\code{"presencia_ausencia"}, \code{"poisson"}, \code{"binomial"},
\code{"negativa_binomial"} o \code{"binomial_negativa"}.}

\item{diagnosticos}{Valor logico. Si es \code{TRUE}, genera diagnosticos de residuos simulados con \code{DHARMa}.}
}
\value{
Un objeto unificado S3 de clase \code{easy_model}.
}
\description{
Esta funcion ajusta un modelo lineal mixto generalizado (GLMM) utilizando \code{lme4::glmer}
con la familia correspondiente segun el tipo de respuesta (Poisson,
Binomial o Binomial Negativa).
En el caso de conteos, calcula e informa el ratio de sobredispersion de Pearson.
Devuelve un objeto unificado de clase \code{easy_model}.
}
\examples{
\dontrun{
  datos <- data.frame(
    exito = rbinom(100, 1, 0.5),
    conteos = rpois(100, lambda = 3),
    x = rnorm(100),
    bloque = factor(rep(1:10, each = 10))
  )
  # Para presencia/ausencia (Binomial)
  modelo <- analizar_glmm(datos, exito ~ x, "(1 | bloque)", tipo = "presencia_ausencia")
  summary(modelo)
}
}
