% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analizar_lmm.R
\name{analizar_lmm}
\alias{analizar_lmm}
\title{Ajustar y analizar un Modelo Lineal Mixto (LMM)}
\usage{
analizar_lmm(
  datos,
  formula_fijos,
  aleatorios,
  REML = TRUE,
  diagnosticos = TRUE
)
}
\arguments{
\item{datos}{Un \code{data.frame} que contiene las variables del modelo.}

\item{formula_fijos}{Una fórmula de R o una cadena de caracteres para la parte de efectos fijos (ej. \code{y ~ x1}).}

\item{aleatorios}{Una cadena de caracteres que define la estructura de efectos aleatorios (ej. \code{"(1 | bloque)"}).}

\item{REML}{Valor lógico. Use \code{FALSE} para comparar modelos con
distintos efectos fijos y \code{TRUE} para el modelo final gaussiano.}

\item{diagnosticos}{Valor lógico. Si es \code{TRUE}, muestra gráficos de
diagnóstico de residuos.}
}
\value{
Un objeto unificado S3 de clase \code{easy_model}.
}
\description{
Esta funcion ajusta un modelo lineal mixto (LMM) utilizando \code{lme4::lmer}.
Adicionalmente, genera gráficos de diagnóstico de residuos (Homocedasticidad y Q-Q plot)
en una sola ventana gráfica y devuelve un objeto unificado de clase \code{easy_model}.
}
\examples{
\dontrun{
  datos <- data.frame(
    y = rnorm(100),
    x = rnorm(100),
    bloque = factor(rep(1:10, each = 10))
  )
  modelo <- analizar_lmm(datos, y ~ x, "(1 | bloque)")
  summary(modelo)
}
}
