% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_modelos.R
\name{calcular_sobredispersion}
\alias{calcular_sobredispersion}
\title{Calcular sobredispersion de Pearson}
\usage{
calcular_sobredispersion(modelo)
}
\arguments{
\item{modelo}{Modelo ajustado.}
}
\value{
data.frame con chi2, gl y ratio.
}
\description{
Calcular sobredispersion de Pearson
}
\keyword{internal}
