% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_modelos.R
\name{construir_formula_mixta}
\alias{construir_formula_mixta}
\title{Construir formula mixta}
\usage{
construir_formula_mixta(formula_fijos, aleatorios)
}
\arguments{
\item{formula_fijos}{Formula de efectos fijos o caracter.}

\item{aleatorios}{Caracter con la definicion de efectos aleatorios (ej. "(1 | bloque)").}
}
\value{
Un objeto de clase \code{formula}.
}
\description{
Concatena la formula de efectos fijos con la estructura de efectos aleatorios.
}
\keyword{internal}
