% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_internos.R
\name{crear_easy_model}
\alias{crear_easy_model}
\title{Crear un objeto unificado de clase easy_model}
\usage{
crear_easy_model(modelo, tipo_modelo, datos, custom_class = NULL)
}
\arguments{
\item{modelo}{Modelo nativo ajustado (ej. lm, glm, merMod, glmmTMB).}

\item{tipo_modelo}{Cadena de caracteres que define el tipo de modelo (ej. "LM", "GLM", "LMM", "GLMM", "RCBD", "Split-Plot").}

\item{datos}{El \code{data.frame} de datos original utilizado en el ajuste.}

\item{custom_class}{Clase S3 adicional para anteponer a "easy_model" (ej. "easy_splitplot").}
}
\value{
Un objeto de clase S3 \code{easy_model} (y opcionalmente la subclase especificada).
}
\description{
Esta funcion interna construye un objeto S3 de clase \code{easy_model} a partir
de un modelo nativo ajustado. Extrae de forma automatica metodos, formulas,
familias y links utilizando el paquete \code{insight}.
}
\keyword{internal}
