% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_modelos.R
\name{detectar_columna_contraste}
\alias{detectar_columna_contraste}
\title{Detectar columna de contraste en salida de emmeans}
\usage{
detectar_columna_contraste(datos)
}
\arguments{
\item{datos}{data.frame devuelto por summary(contrast).}
}
\value{
Nombre de la columna encontrada.
}
\description{
Detectar columna de contraste en salida de emmeans
}
\keyword{internal}
