% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_modelos.R
\name{detectar_columna_prediccion}
\alias{detectar_columna_prediccion}
\title{Detectar columna de prediccion en salida de emmeans}
\usage{
detectar_columna_prediccion(datos)
}
\arguments{
\item{datos}{data.frame devuelto por summary(emm).}
}
\value{
Nombre de la columna encontrada.
}
\description{
Detectar columna de prediccion en salida de emmeans
}
\keyword{internal}
