% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_modelos.R
\name{detectar_intervalos}
\alias{detectar_intervalos}
\title{Detectar columnas de intervalos de confianza}
\usage{
detectar_intervalos(datos)
}
\arguments{
\item{datos}{data.frame.}
}
\value{
Lista con 'inferior' y 'superior' o NULL.
}
\description{
Detectar columnas de intervalos de confianza
}
\keyword{internal}
