% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_internos.R
\name{extraer_modelo}
\alias{extraer_modelo}
\title{Extraer el modelo estadistico subyacente}
\usage{
extraer_modelo(x)
}
\arguments{
\item{x}{Un objeto de clase \code{easy_model} o un modelo nativo (lm, glm, merMod, glmmTMB, etc.).}
}
\value{
El modelo nativo subyacente.
}
\description{
Esta funcion detecta si el objeto de entrada es de clase \code{easy_model}
y extrae el modelo nativo subyacente. Si ya es un modelo nativo, lo devuelve tal cual.
}
