% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc_emmeans.R
\name{obtener_emmeans}
\alias{obtener_emmeans}
\title{Obtener medias marginales estimadas (EMMeans)}
\usage{
obtener_emmeans(modelo, predictor, by = NULL, tipo_respuesta = "response")
}
\arguments{
\item{modelo}{Un objeto de clase \code{easy_model} o un modelo nativo (lm, glm, merMod, glmmTMB, etc.).}

\item{predictor}{Vector de caracteres con el nombre de los predictores a evaluar (ej. \code{"tratamiento"} o \code{c("tratamiento", "dosis")}).}

\item{by}{Vector de caracteres opcional con las variables condicionantes (ej. \code{"bloque"}).}

\item{tipo_respuesta}{Escala del resultado. Use \code{"response"} para la escala de respuesta biologica (probabilidad, tasa, conteos) o \code{"link"} para la escala lineal del predictor.}
}
\value{
Un objeto de clase \code{emmGrid}.
}
\description{
Esta funcion calcula las medias marginales estimadas (EMMeans) para uno o
mas predictores a partir de un modelo unificado de clase \code{easy_model}
o un modelo nativo.
}
\examples{
\dontrun{
  modelo <- analizar_lm(iris, Sepal.Length ~ Species, diagnosticos = FALSE)
  obtener_emmeans(modelo, "Species")
}
}
