% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_modelos.R
\name{obtener_familia_modelo}
\alias{obtener_familia_modelo}
\title{Obtener familia de distribucion del modelo}
\usage{
obtener_familia_modelo(modelo)
}
\arguments{
\item{modelo}{Modelo ajustado.}
}
\value{
Objeto de familia o NULL.
}
\description{
Obtener familia de distribucion del modelo
}
\keyword{internal}
