% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metodos_s3.R
\name{print.easy_splitplot}
\alias{print.easy_splitplot}
\title{Imprimir un objeto easy_splitplot}
\usage{
\method{print}{easy_splitplot}(x, ...)
}
\arguments{
\item{x}{Un objeto de clase \code{easy_splitplot}.}

\item{...}{Argumentos adicionales no utilizados.}
}
\value{
Devuelve el objeto \code{x} de forma invisible.
}
\description{
Muestra un resumen estetico especifico para modelos de parcelas divididas (Split-Plot).
}
