% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metodos_s3.R
\name{print.summary_easy_model}
\alias{print.summary_easy_model}
\title{Imprimir resumen de easy_model}
\usage{
\method{print}{summary_easy_model}(x, ...)
}
\arguments{
\item{x}{Objeto de clase \code{summary_easy_model}.}

\item{...}{Argumentos adicionales.}
}
\value{
Devuelve \code{x} de forma invisible.
}
\description{
Imprimir resumen de easy_model
}
