% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_modelos.R
\name{validar_predictor_modelo}
\alias{validar_predictor_modelo}
\title{Validar existencia de un predictor en el modelo}
\usage{
validar_predictor_modelo(modelo, predictor)
}
\arguments{
\item{modelo}{Modelo ajustado o easy_model.}

\item{predictor}{Nombre del predictor.}
}
\value{
TRUE invisible o error.
}
\description{
Validar existencia de un predictor en el modelo
}
\keyword{internal}
