% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analizar_glm.R
\name{analizar_glm}
\alias{analizar_glm}
\title{Ajustar y analizar un modelo lineal generalizado}
\usage{
analizar_glm(
  datos,
  formula,
  familia = c("gaussian", "binomial", "poisson", "quasibinomial", "quasipoisson",
    "negativa_binomial"),
  diagnostico_dharma = TRUE
)
}
\arguments{
\item{datos}{Un \code{data.frame} con las variables del modelo.}

\item{formula}{Una formula de R para el GLM.}

\item{familia}{Familia del modelo. Puede ser \code{"gaussian"},
\code{"binomial"}, \code{"poisson"}, \code{"quasibinomial"},
\code{"quasipoisson"} o \code{"negativa_binomial"}.}

\item{diagnostico_dharma}{Valor logico. Si es \code{TRUE}, genera
diagnosticos de residuos simulados con \code{DHARMa}.}
}
\value{
Un objeto de clase \code{glm} o \code{negbin}.
}
\description{
Ajusta modelos GLM gaussianos, binomiales, Poisson, cuasi-Poisson,
cuasi-binomiales o binomial negativa. Para modelos de conteo calcula
sobredispersion; para binomiales imprime odds ratios; y puede ejecutar
diagnosticos con \code{DHARMa}.
}
