% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analizar_lmm.R
\name{analizar_lmm}
\alias{analizar_lmm}
\title{Ajustar y analizar un Modelo Lineal Mixto (LMM)}
\usage{
analizar_lmm(
  datos,
  formula_fijos,
  aleatorios,
  REML = TRUE,
  diagnosticos = TRUE
)
}
\arguments{
\item{datos}{Un \code{data.frame} que contiene las variables del modelo.}

\item{formula_fijos}{Una formula de R o una cadena de caracteres para la parte de efectos fijos (ej. \code{y ~ x1}).}

\item{aleatorios}{Una cadena de caracteres que define la estructura de efectos aleatorios (ej. \code{"(1 | bloque)"}).}

\item{REML}{Valor logico. Use \code{FALSE} para comparar modelos con
distintos efectos fijos y \code{TRUE} para el modelo final gaussiano.}

\item{diagnosticos}{Valor logico. Si es \code{TRUE}, muestra graficos de
diagnostico de residuos.}
}
\value{
Un objeto de la clase \code{merMod} correspondiente al modelo ajustado.
}
\description{
Esta funcion ajusta un modelo lineal mixto (LMM) utilizando \code{lme4::lmer}.
Adicionalmente, genera graficos de diagnostico de residuos (Homocedasticidad y Q-Q plot)
en una sola ventana grafica, imprime el resumen (\code{summary}) en la consola y
devuelve el objeto del modelo ajustado para su uso en analisis posteriores (ej. post-hoc).
}
\examples{
\dontrun{
  datos <- data.frame(
    y = rnorm(100),
    x = rnorm(100),
    bloque = factor(rep(1:10, each = 10))
  )
  analizar_lmm(datos, y ~ x, "(1 | bloque)")
}
}
