% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metodos_s3.R
\name{print.easy_model}
\alias{print.easy_model}
\title{Imprimir un objeto easy_model}
\usage{
\method{print}{easy_model}(x, ...)
}
\arguments{
\item{x}{Un objeto de clase \code{easy_model}.}

\item{...}{Argumentos adicionales no utilizados.}
}
\description{
Muestra un resumen estetico y corto del modelo ajustado utilizando
cajas y colores del paquete \code{cli}.
}
