% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metodos_s3.R
\name{summary.easy_model}
\alias{summary.easy_model}
\title{Resumen de un objeto easy_model}
\usage{
\method{summary}{easy_model}(object, ...)
}
\arguments{
\item{object}{Un objeto de clase \code{easy_model}.}

\item{...}{Argumentos adicionales.}
}
\value{
Una lista con el resumen del modelo y la tabla ANOVA.
}
\description{
Devuelve el resumen estadistico detallado del modelo nativo
y la tabla de ANOVA (usando \code{car::Anova} tipo III por defecto para modelos mixtos/split-plot).
}
